IMP - Refactor ATK skill input data utils
