+++ src/components/elements/gameData/skillAtk/in/main.test.tsx
@@ -6,7 +6,7 @@ import userEvent from '@testing-library/user-event';
 import {renderReact} from '../../../../../../test/render/main';
 import {translation as translationEN} from '../../../../../i18n/translations/en/translation';
 import {AttackingSkillInput} from './main';
-import {generateInputData, overwriteInputData} from './utils';
+import {generateInputData, overwriteInputData} from './utils/inputData';
 
 
 describe('ATK skill input', () => {
