+++ src/components/elements/gameData/skillAtk/in/sections/punisher.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import {useI18n} from '../../../../../../i18n/hook';
 import {InputPanel} from '../../../../input/main';
 import {SectionProps} from '../types';
-import {overwriteInputData} from '../utils';
+import {overwriteInputData} from '../utils/inputData';
 
 
 export const SectionPunisher = ({inputData, setInputData}: SectionProps) => {
