+++ src/utils/game/damage.test.ts
@@ -1,7 +1,7 @@
 import {generateAttackingSkillEntry} from '../../../test/data/mock/skill';
 import {AttackingSkillData} from '../../api-def/resources';
 import {InputData} from '../../components/elements/gameData/skillAtk/in/types';
-import {generateInputData} from '../../components/elements/gameData/skillAtk/in/utils';
+import {generateInputData} from '../../components/elements/gameData/skillAtk/in/utils/inputData';
 import {ConditionCodes} from '../../const/gameData';
 import {calculateDamage, CalculateDamageReturn} from './damage';
